// apps/web/src/modules/companies/types.ts
export interface Company {
  id: string
  cnpj: string
  legalName: string
  tradeName: string
  registrationStatus: string
  openingDate: string
  legalNature: string
  primaryCnae: string
  cep: string
  address: string
  number: string
  complement: string
  neighborhood: string
  city: string
  state: string
  phone: string
  mobile: string
  institutionalEmail: string
  processOwner: string
  processOwnerRole: string
}